move: add model dir
